requirements.txt
discord.py==1.7.3
openai==0.10.1
argparse
parse